+++ swarm/src/handler/multi.rs
@@ -86,7 +86,6 @@ where
         Ok(m)
     }
 
-    #[expect(deprecated)] // TODO: Remove when {In, Out}boundOpenInfo is fully removed.
     fn on_listen_upgrade_error(
         &mut self,
         ListenUpgradeError {
@@ -108,7 +107,6 @@ where
     }
 }
 
-#[expect(deprecated)] // TODO: Remove when {In, Out}boundOpenInfo is fully removed.
 impl<K, H> ConnectionHandler for MultiHandler<K, H>
 where
     K: Clone + Debug + Hash + Eq + Send + 'static,