+++ swarm/src/handler/select.rs
@@ -152,7 +152,6 @@ where
     TProto1: ConnectionHandler,
     TProto2: ConnectionHandler,
 {
-    #[expect(deprecated)] // TODO: Remove when {In, Out}boundOpenInfo is fully removed.
     fn on_listen_upgrade_error(
         &mut self,
         ListenUpgradeError {
@@ -182,7 +181,6 @@ where
     }
 }
 
-#[expect(deprecated)] // TODO: Remove when {In, Out}boundOpenInfo is fully removed.
 impl<TProto1, TProto2> ConnectionHandler for ConnectionHandlerSelect<TProto1, TProto2>
 where
     TProto1: ConnectionHandler,